test(contracts): share file-walking helpers between the shipped-surface scans
ALWAYS_EXCLUDED_DIR_NAMES, walkByExt, and flatDir move to
test/contracts/_walk-helpers.mjs (underscore prefix keeps the module out of
the test runner glob, like _rule-helpers.mjs). Both shipped-surface contract
tests import them; the scripts test's inline commands/ and agents/ readdir
blocks become flatDir calls. walkFiles stays local to the scripts test (it
feeds the packed-set expansion with different semantics). Behavior is
byte-identical: same surface roots, extensions, and pinned files.
